+++ changelog.md
@@ -1,6 +1,6 @@
 # goNORTH Changelog
 
-# Unreleased
+# 9.1.0
 
 * 2024-12-12 Added containerHasItem function, made playerHasItem and Item.removeItem functions more robust.
 * 2024-12-10 Scroll panel fades out at the bottom to indicate there's more to scroll through.